move document endpoints to its own file & use binding to validate form
