+++ backend/internal/v1/v1_projects/types.go
@@ -1,8 +1,9 @@
 package v1_projects
 
 import (
-	"KonferCA/SPUR/internal/interfaces"
 	"KonferCA/SPUR/db"
+	"KonferCA/SPUR/internal/interfaces"
+	"mime/multipart"
 )
 
 type Handler struct {
@@ -15,12 +16,12 @@ type CreateProjectRequest struct {
 }
 
 type ProjectResponse struct {
-	ID          string         `json:"id"`
-	Title       string         `json:"title"`
-	Description string         `json:"description"`
+	ID          string           `json:"id"`
+	Title       string           `json:"title"`
+	Description string           `json:"description"`
 	Status      db.ProjectStatus `json:"status"`
-	CreatedAt   int64         `json:"created_at"`
-	UpdatedAt   int64         `json:"updated_at"`
+	CreatedAt   int64            `json:"created_at"`
+	UpdatedAt   int64            `json:"updated_at"`
 }
 
 type ProjectAnswerResponse struct {
@@ -37,8 +38,11 @@ type PatchAnswerRequest struct {
 }
 
 type UploadDocumentRequest struct {
-	Name    string `json:"name" validate:"required"`
-	Section string `json:"section" validate:"required"`
+	File       *multipart.FileHeader `form:"file"`
+	QuestionID string                `form:"question_id" validate:"required,uuid"`
+	Name       string                `form:"name" validate:"required"`
+	Section    string                `form:"section" validate:"required"`
+	SubSection string                `form:"sub_section" validate:"required"`
 }
 
 type DocumentResponse struct {
@@ -66,7 +70,7 @@ type AnswerSubmission struct {
 }
 
 type SubmitProjectResponse struct {
-	Message string         `json:"message"`
+	Message string           `json:"message"`
 	Status  db.ProjectStatus `json:"status"`
 }
 
@@ -97,7 +101,7 @@ type CommentResponse struct {
 	TargetID    string `json:"target_id"`
 	Comment     string `json:"comment"`
 	CommenterID string `json:"commenter_id"`
-	Resolved   bool   `json:"resolved"`
+	Resolved    bool   `json:"resolved"`
 	CreatedAt   int64  `json:"created_at"`
 	UpdatedAt   int64  `json:"updated_at"`
 }
